// src/common/config/environment.enum.ts
export enum Environment {
  PORT = 'PORT',
  RDS_NAME = 'RDS_NAME',
  RDS_DB = 'RDS_DB',
  RDS_HOST = 'RDS_HOST',
  RDS_PORT = 'RDS_PORT',
  RDS_USERNAME = 'RDS_USERNAME',
  RDS_PASSWORD = 'RDS_PASSWORD',
  TYPEORM_LOGGING = 'TYPEORM_LOGGING',
  TYPEORM_SYNCHRONIZE = 'TYPEORM_SYNCHRONIZE',
  CORS_LIST = 'CORS_LIST',
  JWT_SECRET = 'JWT_SECRET',
  JWT_EXPIRED_TIME = 'JWT_EXPIRED_TIME',
}